Refactor ConnectionController and ConnectionControllerGrpc
