+++ src/Examples/ConnHiddenCalc/ConnectionAutomationApp/App.config
@@ -11,7 +11,7 @@
     <applicationSettings>
         <ConnectionAutomationApp.Properties.Settings>
             <setting name="IdeaStatiCaDir" serializeAs="String">
-                <value>C:\Program Files\IDEA StatiCa\StatiCa 22.1</value>
+                <value>C:\Program Files\IDEA StatiCa\StatiCa 23.0</value>
             </setting>
         </ConnectionAutomationApp.Properties.Settings>
     </applicationSettings>
