+++ src/Examples/ConnHiddenCalc/ConnectionAutomationApp/MainVM.cs
@@ -11,7 +11,6 @@ namespace ConnectionAutomationApp
 	public class MainVM : INotifyPropertyChanged, IDisposable
 	{
 		bool isIdea;
-		bool useGrpcCommunication;
 		readonly string ideaStatiCaDir;
 		string statusMessage;
 		IConnectionController connectionController;
@@ -21,7 +20,6 @@ namespace ConnectionAutomationApp
 
 		public MainVM()
 		{
-			UseGrpcCommunication = true;
 			ideaStatiCaDir = Properties.Settings.Default.IdeaStatiCaDir;
 			if (Directory.Exists(ideaStatiCaDir))
 			{
@@ -57,17 +55,6 @@ namespace ConnectionAutomationApp
 			{
 				isIdea = value;
 				NotifyPropertyChanged("IsIdea");
-			}
-		}
-
-		public bool UseGrpcCommunication
-		{
-			get => useGrpcCommunication;
-
-			set
-			{
-				useGrpcCommunication = value;
-				NotifyPropertyChanged("UseGrpcCommunication");
 			}
 		}
 
@@ -142,14 +129,10 @@ namespace ConnectionAutomationApp
 
 		private void RunIdeaConnection(object obj)
 		{
-			if(UseGrpcCommunication)
-			{
-				this.ConnectionController = IdeaConnectionControllerGrpc.Create(ideaStatiCaDir, new NullLogger());
-			}
-			else
-			{
-				this.ConnectionController = IdeaConnectionController.Create(ideaStatiCaDir);
-			}
+			// IdeaConnectionControllerGrpc is obsolete controller.
+			// ConnectionController = IdeaConnectionControllerGrpc.Create(ideaStatiCaDir, new NullLogger());
+
+			ConnectionController = IdeaConnectionController.Create(ideaStatiCaDir);
 		}
 
 		private bool CanRunIdeaConnection(object arg)
