+++ src/Examples/ConnHiddenCalc/ConnectionAutomationApp/MainWindow.xaml
@@ -1,14 +1,13 @@
-﻿<Window x:Class="ConnectionAutomationApp.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:ConnectionAutomationApp"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="450" Width="800" Closed="Window_Closed">
+﻿<Window
+	x:Class="ConnectionAutomationApp.MainWindow"
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	mc:Ignorable="d"
+	Title="MainWindow" Height="450" Width="800" Closed="Window_Closed">
 	<DockPanel LastChildFill="True">
 		<ToolBar DockPanel.Dock="Top">
-			<CheckBox IsChecked="{Binding UseGrpcCommunication, Mode=TwoWay}">Use GRPC</CheckBox>
 			<Button Command="{Binding RunIdeaConnectionCmd, Mode=TwoWay}">Run IdeaConnection</Button>
 			<Button Command="{Binding OpenProjectCmd, Mode=TwoWay}">Open Project</Button>
 			<Button Command="{Binding CloseProjectCmd, Mode=TwoWay}">Close Project</Button>
