+++ src/Examples/ConnHiddenCalc/ConnectionAutomationApp/Properties/Settings.settings
@@ -3,7 +3,7 @@
   <Profiles />
   <Settings>
     <Setting Name="IdeaStatiCaDir" Type="System.String" Scope="Application">
-      <Value Profile="(Default)">C:\Program Files\IDEA StatiCa\StatiCa 22.1</Value>
+      <Value Profile="(Default)">C:\Program Files\IDEA StatiCa\StatiCa 23.0</Value>
     </Setting>
   </Settings>
 </SettingsFile>
